renamed scrabble.rb to scoring.rb
