+++ lib/scoring.rb
